Initial commit for megaDotter the megabase dotplotter
megaDotter uses STL map to index the positions of sequences to create a
dotplot of large regions. The memory required will depend on sequence
complexity and the length of the sequence, with self-dotplots of 4 MB
sequences usually taking up to 1 GB of memory.

This is unlikely to be close to any sort of optimum given the
performance of programs like Blat / Bowtie, but it's simple to code and
simple to understand, and has been useful for me.

In it's current incarnation megaDotter only allows self-comparisons, as
it was written to visualise the extent of repetitiveness of genomic
regions.